Extract a connection class
